requirements.txt
torch
torchvision
pandas
numpy
websocket-client
requests
openpyxl
httpx